# [Enhancement] accumulate chunk into full one in connector chunk source