--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -754,7 +754,7 @@ CONF_Int32(io_coalesce_read_max_buffer_size, "8388608");
 CONF_Int32(io_coalesce_read_max_distance_size, "1048576");
 
 CONF_Int32(connector_io_tasks_per_scan_operator, "16");
-CONF_Bool(connector_chunk_source_accumulate_enable, "true");
+CONF_Bool(connector_chunk_source_accumulate_chunk_enable, "true");
 
 // Enable output trace logs in aws-sdk-cpp for diagnosis purpose.
 // Once logging is enabled in your application, the SDK will generate log files in your current working directory