--- a/be/src/exec/pipeline/chunk_accumulate_operator.cpp
+++ b/be/src/exec/pipeline/chunk_accumulate_operator.cpp
@@ -9,32 +9,17 @@ namespace starrocks {
 namespace pipeline {
 
 Status ChunkAccumulateOperator::push_chunk(RuntimeState* state, const vectorized::ChunkPtr& chunk) {
-    DCHECK(_out_chunk == nullptr);
-
-    if (_in_chunk == nullptr) {
-        _in_chunk = chunk;
-    } else if (_in_chunk->num_rows() + chunk->num_rows() > state->chunk_size()) {
-        _out_chunk = std::move(_in_chunk);
-        _in_chunk = chunk;
-    } else {
-        _in_chunk->append(*chunk);
-    }
-
-    if (_out_chunk == nullptr && (_in_chunk->num_rows() >= state->chunk_size() * LOW_WATERMARK_ROWS_RATE ||
-                                  _in_chunk->memory_usage() >= LOW_WATERMARK_BYTES)) {
-        _out_chunk = std::move(_in_chunk);
-    }
-
+    _acc.push(chunk);
     return Status::OK();
 }
 
 StatusOr<vectorized::ChunkPtr> ChunkAccumulateOperator::pull_chunk(RuntimeState*) {
     // If there isn't more input chunk and _out_chunk has been outputted, output _in_chunk this time.
-    if (_is_finished && _out_chunk == nullptr) {
-        return std::move(_in_chunk);
+    if (_is_finished && _acc.output_chunk() == nullptr) {
+        return std::move(_acc.staging_chunk());
     }
 
-    return std::move(_out_chunk);
+    return std::move(_acc.output_chunk());
 }
 
 Status ChunkAccumulateOperator::set_finishing(RuntimeState* state) {
@@ -45,8 +30,7 @@ Status ChunkAccumulateOperator::set_finishing(RuntimeState* state) {
 
 Status ChunkAccumulateOperator::set_finished(RuntimeState*) {
     _is_finished = true;
-    _in_chunk.reset();
-    _out_chunk.reset();
+    _acc.reset();
 
     return Status::OK();
 }