--- a/be/src/exec/pipeline/chunk_accumulate_operator.cpp
+++ b/be/src/exec/pipeline/chunk_accumulate_operator.cpp
@@ -37,8 +37,7 @@ Status ChunkAccumulateOperator::set_finished(RuntimeState*) {
 
 Status ChunkAccumulateOperator::reset_state(RuntimeState* state, const std::vector<ChunkPtr>& refill_chunks) {
     _is_finished = false;
-    _in_chunk = nullptr;
-    _out_chunk = nullptr;
+    _acc.reset();
 
     return Status::OK();
 }