--- a/be/src/exec/pipeline/chunk_accumulate_operator.h
+++ b/be/src/exec/pipeline/chunk_accumulate_operator.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "exec/pipeline/operator.h"
+#include "storage/chunk_helper.h"
 
 namespace starrocks {
 
@@ -21,9 +22,13 @@ public:
     Status push_chunk(RuntimeState* state, const vectorized::ChunkPtr& chunk) override;
     StatusOr<vectorized::ChunkPtr> pull_chunk(RuntimeState* state) override;
 
-    bool has_output() const override { return _out_chunk != nullptr || (_is_finished && _in_chunk != nullptr); }
-    bool need_input() const override { return !_is_finished && _out_chunk == nullptr; }
-    bool is_finished() const override { return _is_finished && _in_chunk == nullptr && _out_chunk == nullptr; }
+    bool has_output() const override {
+        return _acc.output_chunk() != nullptr || (_is_finished && _acc.staging_chunk() != nullptr);
+    }
+    bool need_input() const override { return !_is_finished && _acc.output_chunk() == nullptr; }
+    bool is_finished() const override {
+        return _is_finished && _acc.staging_chunk() == nullptr && _acc.output_chunk() == nullptr;
+    }
 
     Status set_finishing(RuntimeState* state) override;
     Status set_finished(RuntimeState* state) override;
@@ -31,12 +36,8 @@ public:
     Status reset_state(RuntimeState* state, const std::vector<ChunkPtr>& refill_chunks) override;
 
 private:
-    static constexpr double LOW_WATERMARK_ROWS_RATE = 0.75;          // 0.75 * chunk_size
-    static constexpr size_t LOW_WATERMARK_BYTES = 256 * 1024 * 1024; // 256MB.
-
     bool _is_finished = false;
-    vectorized::ChunkPtr _in_chunk = nullptr;
-    vectorized::ChunkPtr _out_chunk = nullptr;
+    ChunkPipelineAccumulator _acc;
 };
 
 class ChunkAccumulateOperatorFactory final : public OperatorFactory {