--- a/be/src/exec/pipeline/scan/connector_scan_operator.h
+++ b/be/src/exec/pipeline/scan/connector_scan_operator.h
@@ -9,7 +9,6 @@
 #include "exec/workgroup/work_group_fwd.h"
 #include "storage/chunk_helper.h"
 
-
 namespace starrocks {
 
 class ScanNode;