--- a/be/src/storage/chunk_helper.h
+++ b/be/src/storage/chunk_helper.h
@@ -91,4 +91,23 @@ private:
     std::deque<vectorized::ChunkPtr> _output;
 };
 
+class ChunkPipelineAccumulator {
+public:
+    ChunkPipelineAccumulator() = default;
+    void set_max_size(size_t max_size) { _max_size = max_size; }
+    void push(const vectorized::ChunkPtr& chunk);
+    vectorized::ChunkPtr& output_chunk() { return _out_chunk; }
+    vectorized::ChunkPtr& staging_chunk() { return _in_chunk; }
+    const vectorized::ChunkPtr& output_chunk() const { return _out_chunk; }
+    const vectorized::ChunkPtr& staging_chunk() const { return _in_chunk; }
+    void reset();
+
+private:
+    static constexpr double LOW_WATERMARK_ROWS_RATE = 0.75;          // 0.75 * chunk_size
+    static constexpr size_t LOW_WATERMARK_BYTES = 256 * 1024 * 1024; // 256MB.
+    vectorized::ChunkPtr _in_chunk = nullptr;
+    vectorized::ChunkPtr _out_chunk = nullptr;
+    size_t _max_size = 4096;
+};
+
 } // namespace starrocks