## What type of PR is this：
- [ ] BugFix
- [ ] Feature
- [X] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

ssb-100g with block cache, optimized version can save almost 1000ms. 



  | ssbflat(before) | after | ssb(before) | after
-- | -- | -- | -- | --
Q01 | 444 | 350 | 683 | 612
Q02 | 419 | 328 | 645 | 574
Q03 | 460 | 414 | 588 | 559
Q04 | 628 | 445 | 878 | 845
Q05 | 478 | 360 | 729 | 723
Q06 | 335 | 276 | 699 | 693
Q07 | 920 | 666 | 1372 | 1396
Q08 | 817 | 601 | 823 | 807
Q09 | 298 | 275 | 741 | 737
Q10 | 219 | 222 | 762 | 667
Q11 | 861 | 608 | 1573 | 1577
Q12 | 1027 | 748 | 1310 | 1168
Q13 | 958 | 738 | 1210 | 1095
SUM | 7864 | 6031 | 12013 | 11453

tpch-100g with block cache.

  | tpch(before) | after
-- | -- | --
Q01 | 3654 | 3655
Q02 | 513 | 478
Q03 | 2425 | 2293
Q04 | 1150 | 1069
Q05 | 1904 | 1874
Q06 | 896 | 774
Q07 | 1726 | 1682
Q08 | 1577 | 1530
Q09 | 4756 | 4699
Q10 | 2764 | 2590
Q11 | 380 | 379
Q12 | 1366 | 1204
Q13 | 2761 | 2796
Q14 | 1119 | 971
Q15 | 996 | 845
Q16 | 3535 | 3615
Q17 | 1869 | 1775
Q18 | 10364 | 10369
Q19 | 1173 | 1000
Q20 | 1220 | 1181
Q21 | 3513 | 3509
Q22 | 678 | 691
SUM | 50339 | 48979





## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] I have added user document for my new feature or new function