@wanpengfei-git 2022-10-18
run starrocks_admit_test