@wanpengfei-git 2022-10-20
run starrocks_admit_test