@wanpengfei-git 2022-10-21
run starrocks_admit_test