@dirtysalt 2022-10-21
run starrocks_be_unittest