@dirtysalt 2022-10-21
run starrocks_clang-format