@wanpengfei-git 2022-10-21
starrocks_be_unittest succeeded.