@wanpengfei-git 2022-10-22
starrocks_be_unittest succeeded.