@wanpengfei-git 2022-10-23
run starrocks_admit_test