@wanpengfei-git 2022-10-23
starrocks_be_unittest succeeded.